Revert of Various cleanups in crypto code discovered while syncing recent changes. (https://codereview.chromium.org/217653003/)
Reason for revert:
Seems to be related to a tree-closing failure.

http://build.chromium.org/p/chromium.memory/builders/Linux%20ASan%2BLSan%20Tests%20%281%29/builds/984

See bug for more details.

BUG=358919

Original issue's description:
> Various cleanups in crypto code discovered while syncing recent changes.
> 
> Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=260979

TBR=rtenneti@chromium.org,rch@chromium.org
NOTREECHECKS=true
NOTRY=true

Review URL: https://codereview.chromium.org/212223009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@261017 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ef65c323668aed3b9fb880441084dd2807fa5d67

+++ net/quic/crypto/quic_crypto_client_config.cc
@@ -711,7 +711,7 @@ void QuicCryptoClientConfig::PopulateFromCanonicalConfig(
     const QuicServerId& server_id,
     CachedState* server_state) {
   DCHECK(server_state->IsEmpty());
-  size_t i = 0;
+  unsigned i = 0;
   for (; i < canoncial_suffixes_.size(); ++i) {
     if (EndsWith(server_id.host(), canoncial_suffixes_[i], false)) {
       break;
